Removed redundant and obsolete code
I've done a bit of spring cleaning in the codebase removing unused files, old obsolete code, etc.
