+++ CMakeLists.txt
@@ -86,12 +86,9 @@ configure_file(
 add_library(XIAfuncs
         ${CMAKE_CURRENT_SOURCE_DIR}/src/Format/event.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/src/Tasks/Buffer.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/src/Tasks/BufferSE.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/src/Tasks/Calibrator.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/src/Tasks/Sort.cpp
-        #${CMAKE_CURRENT_SOURCE_DIR}/src/Tasks/MTSort.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/src/Tasks/Splitter.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/src/Tasks/SplitterSE.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/src/Tasks/Trigger.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/src/Tasks/XIAReader.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/src/RootInterface/TTreeManager.cpp
