+++ include/Tasks/Splitter.h
@@ -2,8 +2,8 @@
 // Created by Vetle Wegner Ingeberg on 16/04/2021.
 //
 
-#ifndef TDR2TREE_SPLITTER_H
-#define TDR2TREE_SPLITTER_H
+#ifndef SPLITTERSE_H
+#define SPLITTERSE_H
 
 #include <Tasks.h>
 #include <Queue.h>
@@ -13,7 +13,7 @@ namespace Task {
     class Splitter : public Base
     {
     private:
-        EventQueue_t &input_queue;
+        EntryQueue_t &input_queue;
         MCEventQueue_t output_queue;
         const double gap;
         //std::vector<word_t> buffer;
@@ -22,7 +22,7 @@ namespace Task {
 
     public:
 
-        Splitter(EventQueue_t &input, const double &time_gap = 1500., const size_t &cap = 65536);
+        Splitter(EntryQueue_t &input, const double &time_gap = 1500., const size_t &cap = 65536);
         MCEventQueue_t &GetQueue(){ return output_queue; }
         void Run() override;
 
@@ -30,4 +30,4 @@ namespace Task {
 
 }
 
-#endif //TDR2TREE_SPLITTER_H
+#endif //SPLITTERSE_H
